requirements.txt
Pillow==11.1.0
imageio==2.37.0
matplotlib==3.10.1
torch==2.6.0+cu118
torchvision==0.21.0+cu118
wandb==0.19.8
tqdm==4.67.1
wandbx3